removed some composables and componentized the personenkontextCreation
